+++ src/kspdg/tests/ksp_ingame_tests/test_lbg1_i2.py
@@ -17,7 +17,7 @@ import kspdg.utils.constants as C
 from kspdg.lbg1.lg0_envs import LBG1_LG0_I2_Env
 from kspdg.lbg1.lg1_envs import LBG1_LG1_I2_Env
 from kspdg.lbg1.lg2_envs import LBG1_LG2_I2_Env
-from kspdg import LBG1_LG3_I2_V1, LBG1_LG4_I2_V1
+from kspdg import LBG1_LG3_I2_V1, LBG1_LG4_I2_V1, LBG1_LG5_I2_V1
 
 @pytest.fixture
 def lbg1_lg0_i2_env():
@@ -59,6 +59,14 @@ def lbg1_lg4_i2_env():
     yield env
     env.close()
 
+@pytest.fixture
+def lbg1_lg5_i2_env():
+    """setup and teardown of the LBG1_LG5_I2_Env object connected to kRPC server"""
+    env = LBG1_LG5_I2_V1()
+    env.reset()
+    yield env
+    env.close()
+
 def test_smoketest_lg0(lbg1_lg0_i2_env):
     """Ensure no errors are thrown from starting LG0 environment"""
     pass
@@ -77,6 +85,10 @@ def test_smoketest_lg3(lbg1_lg3_i2_env):
 
 def test_smoketest_lg4(lbg1_lg4_i2_env):
     """Ensure no errors are thrown from starting LG4 environment"""
+    pass
+
+def test_smoketest_lg5(lbg1_lg5_i2_env):
+    """Ensure no errors are thrown from starting LG5 environment"""
     pass
 
 def test_observation_0(lbg1_lg0_i2_env):